making the undo/redo page look nice loading app version in title renaming undoredo because existing naming was being yelled at
