+++ app/src/index-prod.html
@@ -3,7 +3,6 @@
 
 <head>
   <meta charset="UTF-8">
-  <title>Getting Started!</title>
   <base href="app://rse">
 </head>
