+++ app/src/index.html
@@ -3,7 +3,6 @@
 
 <head>
   <meta charset="UTF-8">
-  <title>Getting Started!</title>
 </head>
 
 <body>
