--- app/src/pages/undo-redo/undo-redo.css
@@ -1,11 +1,0 @@
-#undoredo {
-    display: grid;
-    height: 100%;
-    justify-items: center;
-    grid-template-columns: auto;
-    grid-template-rows: 1fr 10fr 1fr;     
-}
-
-.left {
-    justify-self: start;
-}
