+++ app/src/pages/undoredo/undoredo.css
@@ -1,0 +1,25 @@
+#undoredo {
+    display: grid;
+    height: 100%;
+    justify-items: center;
+    grid-template-columns: auto;
+    grid-template-rows: 40px 2fr 4fr 2fr 1fr 40px;      
+}
+
+.undoredo {
+    font-size: 50px;
+    text-align: center;
+    grid-column: 1;
+    grid-row: 2;
+}
+
+.left {
+    justify-self: start;
+}
+
+.container {
+    max-height: 400px;
+    overflow-y: scroll;
+    background-color: #282c34;
+    color: #6d9cbe;
+}
